+++ R/utils.R
@@ -44,3 +44,34 @@ relish <- function(x, dir = getwd()) {
   }
   gsub(dir, "", x, fixed = TRUE)
 }
+
+#' Utility function for simplifying single-element lists
+#'
+#' @param x A list of length at least one.
+#'
+#' @return If x has length greater than one, it is returned unchanged. If the length is equal to
+#'     one, the first element is returned.
+simplify_single <- function(x) {
+  if (length(x) == 1) {
+    x[[1]]
+  } else {
+    x
+  }
+}
+
+#' Get value of original option or of newer replacement version
+#'
+#' This funciton exists to handle deprecated options
+#'
+#' @param original A string. Name of the original (pre 0.3.0) option.
+#' @param replacement A string. Name of the new (since 0.3.0) option.
+#'
+#' @keywords internal
+get_opt <- function(original, replacement) {
+  opt <- getOption(original)
+  if (is.null(opt)) {
+    return(getOption(replacement))
+  } else {
+    return(opt)
+  }
+}